+++ sound-mind-studio/tests/test_main_window.h
@@ -31,6 +31,9 @@ private slots:
     void exportTopmostLayerVideoAsyncCompletesSuccessfullyAndHidesTheCancelButton();
     void cancelVideoExportStopsItAndDeletesThePartialFile();
     void exportTopmostLayerVideoAsyncDoesNothingWhileAlreadyRunning();
+    void exportTopmostLayerAudioAsyncCompletesSuccessfullyAndHidesTheCancelButton();
+    void cancelAudioExportStopsItAndDeletesThePartialFile();
+    void exportTopmostLayerAudioAsyncDoesNothingWhileAVideoExportIsRunning();
     void importAudioFileShowsProgressThenCompletionInTheStatusBar();
     void poolTopmostLayerNowShowsProgressThenCompletionInTheStatusBar();
     void exportTopmostLayerAudioNowShowsProgressThenCompletionInTheStatusBar();